chore: move schema
